--- a/packages/ubuntu_desktop_installer/test/active_directory/test_active_directory.dart
+++ b/packages/ubuntu_desktop_installer/test/active_directory/test_active_directory.dart
@@ -5,12 +5,12 @@ import 'package:mockito/mockito.dart';
 import 'package:ubuntu_desktop_installer/pages/active_directory/active_directory_model.dart';
 import 'package:ubuntu_desktop_installer/pages/active_directory/active_directory_page.dart';
 import 'package:ubuntu_desktop_installer/services.dart';
-import 'package:ubuntu_wizard/utils.dart';
 
 import 'test_active_directory.mocks.dart';
+export '../test_utils.dart';
 export 'test_active_directory.mocks.dart';
 
-@GenerateMocks([ActiveDirectoryService])
+@GenerateMocks([ActiveDirectoryModel])
 ActiveDirectoryModel buildActiveDirectoryModel({
   bool? isValid,
   String? domainName,
@@ -37,7 +37,6 @@ ActiveDirectoryModel buildActiveDirectoryModel({
   return model;
 }
 
-@GenerateMocks([ActiveDirectoryModel, UrlLauncher])
 Widget buildActiveDirectoryPage(ActiveDirectoryModel model) {
   return ProviderScope(
     overrides: [activeDirectoryModelProvider.overrideWith((_) => model)],