--- a/packages/ubuntu_desktop_installer/test/filesystem/allocate_disk_space/allocate_disk_space_dialogs_test.dart
+++ b/packages/ubuntu_desktop_installer/test/filesystem/allocate_disk_space/allocate_disk_space_dialogs_test.dart
@@ -14,7 +14,6 @@ import 'package:ubuntu_wizard/utils.dart';
 import 'package:yaru_test/yaru_test.dart';
 import 'package:yaru_widgets/yaru_widgets.dart';
 
-import '../../test_utils.dart';
 import 'test_allocate_disk_space.dart';
 
 void main() {