--- a/packages/ubuntu_desktop_installer/test/filesystem/allocate_disk_space/allocate_disk_space_page_test.dart
+++ b/packages/ubuntu_desktop_installer/test/filesystem/allocate_disk_space/allocate_disk_space_page_test.dart
@@ -12,7 +12,6 @@ import 'package:ubuntu_desktop_installer/services.dart';
 import 'package:yaru_test/yaru_test.dart';
 import 'package:yaru_widgets/yaru_widgets.dart';
 
-import '../../test_utils.dart';
 import 'test_allocate_disk_space.dart';
 
 final testDisks = <Disk>[