--- a/packages/ubuntu_desktop_installer/test/filesystem/allocate_disk_space/test_allocate_disk_space.dart
+++ b/packages/ubuntu_desktop_installer/test/filesystem/allocate_disk_space/test_allocate_disk_space.dart
@@ -3,12 +3,12 @@ import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
 import 'package:subiquity_client/subiquity_client.dart';
 import 'package:ubuntu_desktop_installer/pages/filesystem/allocate_disk_space/allocate_disk_space_model.dart';
-import 'package:ubuntu_desktop_installer/services.dart';
 
 import 'test_allocate_disk_space.mocks.dart';
+export '../../test_utils.dart';
 export 'test_allocate_disk_space.mocks.dart';
 
-@GenerateMocks([AllocateDiskSpaceModel, UdevDeviceInfo, UdevService])
+@GenerateMocks([AllocateDiskSpaceModel])
 AllocateDiskSpaceModel buildAllocateDiskSpaceModel({
   bool? isValid,
   List<Disk>? disks,