--- a/packages/ubuntu_desktop_installer/test/filesystem/installation_type/installation_type_page_test.dart
+++ b/packages/ubuntu_desktop_installer/test/filesystem/installation_type/installation_type_page_test.dart
@@ -8,7 +8,6 @@ import 'package:ubuntu_desktop_installer/pages/filesystem/installation_type/inst
 import 'package:ubuntu_desktop_installer/services.dart';
 import 'package:yaru_test/yaru_test.dart';
 
-import '../../test_utils.dart';
 import 'test_installation_type.dart';
 
 void main() {