Generate and share all service mocks from one central place. This not only removes a nice chunk of duplicate mocks but solves an actual problem in the high-level wizard flow test (#2012) that there are lots of conflicting service mocks when importing all mock models into the same wizard flow test.