@jpnurmi 2023-05-25
P.S. The number of output files goes down from 34 to 25 so I had expected this to speed up the build-runner step but this doesn't seem to be the case... :(

Before:
```
[INFO] Generating build script completed, took 202ms
[INFO] Precompiling build script... completed, took 3.7s
[INFO] Building new asset graph completed, took 414ms
[INFO] Checking for unexpected pre-existing outputs. completed, took 3ms
[INFO] Running build completed, took 15.5s
[INFO] Caching finalized dependency graph completed, took 45ms
[INFO] Succeeded after 15.5s with 34 outputs (96 actions)
```

After:
```
[INFO] Generating build script completed, took 190ms
[INFO] Precompiling build script... completed, took 3.7s
[INFO] Building new asset graph completed, took 466ms
[INFO] Checking for unexpected pre-existing outputs. completed, took 2ms
[INFO] Running build completed, took 15.6s
[INFO] Caching finalized dependency graph completed, took 51ms
[INFO] Succeeded after 15.6s with 25 outputs (96 actions)
```